requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==21.2.0
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.6.0
django-extensions
djangorestframework-simplejwt
